// types/weather-type.ts
export type WeatherType =
  | 'clearday'
  | 'clearnight'
  | 'pcloudyday'
  | 'pcloudynight'
  | 'mcloudyday'
  | 'mcloudynight'
  | 'cloudyday'
  | 'cloudynight'
  | 'humidday'
  | 'humidnight'
  | 'lightrainday'
  | 'lightrainnight'
  | 'oshowerday'
  | 'oshowernight'
  | 'ishowerday'
  | 'ishowernight'
  | 'lightsnowday'
  | 'lightsnownight'
  | 'rainday'
  | 'rainnight'
  | 'snowday'
  | 'snownight'
  | 'rainsnowday'
  | 'rainsnownight';